Create header, add logo to site, begin final design
