+++ app/globals.css
@@ -5,7 +5,7 @@
   --color-secondary: #60A5FA; /* Soft Blue */
   --color-bg-light: #F8FAFC; /* Blue-tinted white */
   --color-section-bg: #E0F2FE; /* Pale Blue */
-  --color-text-primary: #111827; /* Neutral Gray-Black */
+  --color-text-primary: #033b88; /* Neutral Gray-Black */
   --color-text-secondary: #4B5563; /* Medium Gray */
 }
 
@@ -21,7 +21,7 @@ nav.main-nav {
   justify-content: space-between;
   align-items: center;
   padding: 1rem 2rem;
-  background: var(--color-primary);
+  background: #fff;
   border-bottom: 1px solid #ddd;
 }
 
@@ -45,14 +45,17 @@ nav.main-nav {
 }
 
 .nav-links a {
-  color: var(--color-bg-light);
   text-decoration: none;
   font-family: Montserrat, sans-serif;
   font-size: large;
   font-weight: 550;
 }
 
-.nav-links a:hover {
+.main-nav .nav-links a {
+  color: var(--color-text-primary);
+}
+
+.main-nav .nav-links a:hover {
   color: var(--color-accent);
 }
 
@@ -78,12 +81,11 @@ a:hover {
 }
 
 main {
-  padding: 1rem;
+  padding: 0;
 }
 
 section {
   padding: 2rem 1rem;
-  background: var(--color-bg-light);
 }
 
 section.alt-section {
@@ -91,11 +93,11 @@ section.alt-section {
 }
 
 h1, h2, h3 {
-  color: var(--color-text-primary);
+  color: #FFF;
 }
 
 p, label {
-  color: var(--color-text-secondary);
+  color: #FFF;
 }
 
 .gallery {
@@ -114,16 +116,27 @@ form {
 
 button,
 .btn {
-  background: var(--color-accent);
+  background: var(--color-primary);
   color: #fff;
   border: none;
   padding: 0.5rem 1rem;
   cursor: pointer;
+  border-radius: 9999px;
 }
 
 button:hover,
 .btn:hover {
   background: var(--color-secondary);
+  color: #fff;
+}
+
+.btn.secondary {
+  background: transparent;
+  border: 1px solid #fff;
+}
+
+.btn.secondary:hover {
+  background: rgba(255, 255, 255, 0.2);
 }
 
 .card {
@@ -173,4 +186,91 @@ button:hover,
 .hero h1 {
   font-size: 2rem;
   margin-top: 0.5rem;
+}
+
+/* Landing page hero */
+.landing-hero {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+  padding: 0 2rem;
+  color: #fff;
+  overflow: hidden;
+}
+
+.landing-hero .background {
+  position: absolute;
+  inset: 0;
+  z-index: -2;
+}
+
+.landing-hero .background .slide {
+  position: absolute;
+  inset: 0;
+  object-fit: cover;
+  opacity: 0;
+  transition: opacity 1s ease-in-out;
+}
+
+.landing-hero .background .slide.active {
+  opacity: 1;
+}
+
+.landing-hero::after {
+  content: '';
+  position: absolute;
+  inset: 0;
+  background: rgba(0, 0, 0, 0.4);
+  z-index: -1;
+}
+
+.landing-hero .content {
+  max-width: 480px;
+}
+
+.landing-hero .actions {
+  display: flex;
+  gap: 1rem;
+  margin-top: 1rem;
+}
+
+/* Overlay header for landing page */
+.header-overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  z-index: 10;
+}
+
+.header-overlay nav.main-nav {
+  background: transparent;
+  border: none;
+}
+
+.header-overlay .nav-links a:hover{
+    color: #fff;
+}
+
+.header-overlay .nav-links a {
+  color: #fff;
+  position: relative;
+  text-decoration: none;
+}
+
+.header-overlay .nav-links a::after {
+    content: '';
+    position: absolute;
+    left: 0;
+    bottom: -2px; /* Adjust to position the underline */
+    width: 0;
+    height: 2px; /* Thickness of the underline */
+    background-color: #fff;
+    transition: width 0.3s ease;
+}
+
+.header-overlay .nav-links a:hover::after {
+  width: 100%; /* Full width underline on hover */
 }
